pyproject.toml
[tool.poetry]
name = "bot-movie"
version = "0.1.0"
description = ""
authors = ["Dzmitry <isys35@mail.ru>"]
readme = "README.md"
packages = [{include = "bot_movie"}]

[tool.poetry.dependencies]
python = "^3.10"
pre-commit = "^3.0.0"
cinemagoer = "^2022.12.27"
aiogram = "^2.24"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
